fix(ci): make component-model code build & pass on all CI platforms
The component-model merge shipped several cross-platform breaks that the
scratch/BSD/plan9 CI jobs catch but the -short amd64 job did not:

- abi/layout.go: `numCases > math.MaxUint32` overflows int on 32-bit
  (linux/arm, linux/386). Compare as uint64.

- instance/wasi_sockets.go: used syscall.ECONNREFUSED/EADDRINUSE/EADDRNOTAVAIL,
  which don't exist on plan9 (breaking `make check`'s plan9 cross-build). Routed
  through package vars, real on !plan9 and inert sentinels on plan9.

- abi oracle tests read testdata/oracle_*.json via relative os.ReadFile, but the
  scratch/BSD jobs run the compiled test binary from the repo root, so the files
  weren't found (TestFlatOracleLowerFlat etc. failed with "no such file"). Embed
  the golden JSON via //go:embed so it travels with the binary.

- instance socket errno tests do real dials/binds and assert the mapped wasi
  code. On Windows the Winsock errno doesn't match syscall.ECONNREFUSED via
  errors.Is, so those OS-error subtests now skip on Windows (mapping is still
  covered on linux/darwin/BSD).

Verified: abi + instance pass; the abi test binary passes when run from the repo
root (the scratch/BSD invocation); cross-builds clean for plan9/js/wasip1/386/arm.

+++ internal/component/abi/layout.go
@@ -110,7 +110,7 @@ func Size(t binary.TypeDesc, resolve Resolver) (uint32, error) {
 // DiscriminantType returns the core type used to encode a variant discriminant.
 // This mirrors the canonical ABI discriminant_type() function.
 func DiscriminantType(numCases int) string {
-	if numCases <= 0 || numCases > math.MaxUint32 {
+	if numCases <= 0 || uint64(numCases) > math.MaxUint32 {
 		return "" // invalid
 	}
 	// Compute ceil(log2(numCases) / 8)
